+++ src/troute-routing/troute/routing/fast_reach/diffusive.pyx
@@ -6,7 +6,7 @@ from .fortran_wrappers cimport c_diffnw
 
 @cython.boundscheck(False)
 cdef void diffnw(
-        double[::1] timestep_ar_g,
+        float[::1] timestep_ar_g,
         int nts_ql_g,
         int nts_ub_g,
         int nts_db_g,
@@ -15,45 +15,45 @@ cdef void diffnw(
         int nts_da_g,
         int mxncomp_g,
         int nrch_g,
-        double[::1,:] z_ar_g,
-        double[::1,:] bo_ar_g,
-        double[::1,:] traps_ar_g,
-        double[::1,:] tw_ar_g,
-        double[::1,:] twcc_ar_g,
-        double[::1,:] mann_ar_g,
-        double[::1,:] manncc_ar_g,
-        double[::1,:] so_ar_g,
-        double[::1,:] dx_ar_g,
-        double[::1,:] iniq,
+        float[::1,:] z_ar_g,
+        float[::1,:] bo_ar_g,
+        float[::1,:] traps_ar_g,
+        float[::1,:] tw_ar_g,
+        float[::1,:] twcc_ar_g,
+        float[::1,:] mann_ar_g,
+        float[::1,:] manncc_ar_g,
+        float[::1,:] so_ar_g,
+        float[::1,:] dx_ar_g,
+        float[::1,:] iniq,
         int frnw_col,
         int[::1,:] frnw_g,
-        double[::1,:,:] qlat_g,
-        double[::1,:] ubcd_g,
-        double[::1] dbcd_g,
-        double[::1,:] qtrib_g,
+        float[::1,:,:] qlat_g,
+        float[::1,:] ubcd_g,
+        float[::1] dbcd_g,
+        float[::1,:] qtrib_g,
         int paradim,
-        double[::1] para_ar_g,
+        float[::1] para_ar_g,
         int mxnbathy_g,
-        double[::1,:,:] x_bathy_g,
-        double[::1,:,:] z_bathy_g,
-        double[::1,:,:] mann_bathy_g,
+        float[::1,:,:] x_bathy_g,
+        float[::1,:,:] z_bathy_g,
+        float[::1,:,:] mann_bathy_g,
         int[::1,:] size_bathy_g,  
-        double[::1,:] usgs_da_g,
+        float[::1,:] usgs_da_g,
         int[::1] usgs_da_reach_g,
-        double[::1,:] rdx_ar_g,
+        float[::1,:] rdx_ar_g,
         int cwnrow_g,
         int cwncol_g,
-        double[::1,:] crosswalk_g,
-        double[::1,:] z_thalweg_g,
-        double[:,:,:] out_q,
-        double[:,:,:] out_elv,
-        double[:,:,:] out_depth,
+        float[::1,:] crosswalk_g,
+        float[::1,:] z_thalweg_g,
+        float[:,:,:] out_q,
+        float[:,:,:] out_elv,
+        float[:,:,:] out_depth,
 ):
 
     cdef:
-        double[::1,:,:] q_ev_g     = np.empty([ntss_ev_g,mxncomp_g,nrch_g], dtype = np.double, order = 'F')
-        double[::1,:,:] elv_ev_g   = np.empty([ntss_ev_g,mxncomp_g,nrch_g], dtype = np.double, order = 'F')
-        double[::1,:,:] depth_ev_g = np.empty([ntss_ev_g,mxncomp_g,nrch_g], dtype = np.double, order = 'F')
+        float[::1,:,:] q_ev_g     = np.empty([ntss_ev_g,mxncomp_g,nrch_g], dtype = np.float32, order = 'F')
+        float[::1,:,:] elv_ev_g   = np.empty([ntss_ev_g,mxncomp_g,nrch_g], dtype = np.float32, order = 'F')
+        float[::1,:,:] depth_ev_g = np.empty([ntss_ev_g,mxncomp_g,nrch_g], dtype = np.float32, order = 'F')
     
     c_diffnw(
         &timestep_ar_g[0],
@@ -112,7 +112,7 @@ cpdef object compute_diffusive(
 
     # unpack/declare diffusive input variables
     cdef:
-        double[::1] timestep_ar_g = np.asfortranarray(diff_inputs['timestep_ar_g'])
+        float[::1] timestep_ar_g = np.asfortranarray(diff_inputs['timestep_ar_g'].astype(np.float32))
         int nts_ql_g = diff_inputs["nts_ql_g"]
         int nts_ub_g = diff_inputs["nts_ub_g"]
         int nts_db_g = diff_inputs["nts_db_g"]
@@ -121,39 +121,39 @@ cpdef object compute_diffusive(
         int nts_da_g = diff_inputs["nts_da_g"]       
         int mxncomp_g = diff_inputs["mxncomp_g"]
         int nrch_g = diff_inputs["nrch_g"]
-        double[::1,:] z_ar_g = np.asfortranarray(diff_inputs["z_ar_g"])
-        double[::1,:] bo_ar_g = np.asfortranarray(diff_inputs["bo_ar_g"])
-        double[::1,:] traps_ar_g = np.asfortranarray(diff_inputs["traps_ar_g"])
-        double[::1,:] tw_ar_g = np.asfortranarray(diff_inputs["tw_ar_g"])
-        double[::1,:] twcc_ar_g = np.asfortranarray(diff_inputs["twcc_ar_g"])
-        double[::1,:] mann_ar_g = np.asfortranarray(diff_inputs["mann_ar_g"])
-        double[::1,:] manncc_ar_g = np.asfortranarray(diff_inputs["manncc_ar_g"])
-        double[::1,:] so_ar_g = np.asfortranarray(diff_inputs["so_ar_g"])
-        double[::1,:] dx_ar_g = np.asfortranarray(diff_inputs["dx_ar_g"])
-        double[::1,:] iniq = np.asfortranarray(diff_inputs["iniq"])
+        float[::1,:] z_ar_g = np.asfortranarray(diff_inputs["z_ar_g"].astype(np.float32))
+        float[::1,:] bo_ar_g = np.asfortranarray(diff_inputs["bo_ar_g"].astype(np.float32))
+        float[::1,:] traps_ar_g = np.asfortranarray(diff_inputs["traps_ar_g"].astype(np.float32))
+        float[::1,:] tw_ar_g = np.asfortranarray(diff_inputs["tw_ar_g"].astype(np.float32))
+        float[::1,:] twcc_ar_g = np.asfortranarray(diff_inputs["twcc_ar_g"].astype(np.float32))
+        float[::1,:] mann_ar_g = np.asfortranarray(diff_inputs["mann_ar_g"].astype(np.float32))
+        float[::1,:] manncc_ar_g = np.asfortranarray(diff_inputs["manncc_ar_g"].astype(np.float32))
+        float[::1,:] so_ar_g = np.asfortranarray(diff_inputs["so_ar_g"].astype(np.float32))
+        float[::1,:] dx_ar_g = np.asfortranarray(diff_inputs["dx_ar_g"].astype(np.float32))
+        float[::1,:] iniq = np.asfortranarray(diff_inputs["iniq"].astype(np.float32))
         int frnw_col = diff_inputs["frnw_col"]
         int[::1,:] frnw_g = np.asfortranarray(diff_inputs["frnw_g"])
-        double[::1,:,:] qlat_g = np.asfortranarray(diff_inputs["qlat_g"])
-        double[::1,:] ubcd_g = np.asfortranarray(diff_inputs["ubcd_g"])
-        double[::1] dbcd_g = np.asfortranarray(diff_inputs["dbcd_g"])
-        double[::1,:] qtrib_g = np.asfortranarray(diff_inputs["qtrib_g"])
+        float[::1,:,:] qlat_g = np.asfortranarray(diff_inputs["qlat_g"].astype(np.float32))
+        float[::1,:] ubcd_g = np.asfortranarray(diff_inputs["ubcd_g"].astype(np.float32))
+        float[::1] dbcd_g = np.asfortranarray(diff_inputs["dbcd_g"].astype(np.float32))
+        float[::1,:] qtrib_g = np.asfortranarray(diff_inputs["qtrib_g"].astype(np.float32))
         int paradim = diff_inputs['paradim']
-        double[::1] para_ar_g = np.asfortranarray(diff_inputs["para_ar_g"])
+        float[::1] para_ar_g = np.asfortranarray(diff_inputs["para_ar_g"].astype(np.float32))
         int mxnbathy_g = diff_inputs['mxnbathy_g']
-        double[::1,:,:] x_bathy_g = np.asfortranarray(diff_inputs["x_bathy_g"])
-        double[::1,:,:] z_bathy_g = np.asfortranarray(diff_inputs["z_bathy_g"])
-        double[::1,:,:] mann_bathy_g = np.asfortranarray(diff_inputs["mann_bathy_g"])
+        float[::1,:,:] x_bathy_g = np.asfortranarray(diff_inputs["x_bathy_g"].astype(np.float32))
+        float[::1,:,:] z_bathy_g = np.asfortranarray(diff_inputs["z_bathy_g"].astype(np.float32))
+        float[::1,:,:] mann_bathy_g = np.asfortranarray(diff_inputs["mann_bathy_g"].astype(np.float32))
         int[::1,:] size_bathy_g = np.asfortranarray(diff_inputs["size_bathy_g"])    
-        double[::1,:] usgs_da_g = np.asfortranarray(diff_inputs["usgs_da_g"])   
+        float[::1,:] usgs_da_g = np.asfortranarray(diff_inputs["usgs_da_g"].astype(np.float32))   
         int[::1] usgs_da_reach_g = np.asfortranarray(diff_inputs["usgs_da_reach_g"]) 
-        double[::1,:] rdx_ar_g = np.asfortranarray(diff_inputs["rdx_ar_g"])
+        float[::1,:] rdx_ar_g = np.asfortranarray(diff_inputs["rdx_ar_g"].astype(np.float32))
         int cwnrow_g = diff_inputs["cwnrow_g"]
         int cwncol_g = diff_inputs["cwncol_g"]
-        double[::1,:] crosswalk_g = np.asfortranarray(diff_inputs["crosswalk_g"]) 
-        double[::1,:] z_thalweg_g = np.asfortranarray(diff_inputs["z_thalweg_g"])
-        double[:,:,:] out_q = np.empty([ntss_ev_g,mxncomp_g,nrch_g], dtype = np.double)
-        double[:,:,:] out_elv = np.empty([ntss_ev_g,mxncomp_g,nrch_g], dtype = np.double)
-        double[:,:,:] out_depth = np.empty([ntss_ev_g,mxncomp_g,nrch_g], dtype = np.double)
+        float[::1,:] crosswalk_g = np.asfortranarray(diff_inputs["crosswalk_g"].astype(np.float32)) 
+        float[::1,:] z_thalweg_g = np.asfortranarray(diff_inputs["z_thalweg_g"].astype(np.float32))
+        float[:,:,:] out_q = np.empty([ntss_ev_g,mxncomp_g,nrch_g], dtype = np.float32)
+        float[:,:,:] out_elv = np.empty([ntss_ev_g,mxncomp_g,nrch_g], dtype = np.float32)
+        float[:,:,:] out_depth = np.empty([ntss_ev_g,mxncomp_g,nrch_g], dtype = np.float32)
 
     # call diffusive compute kernel
     diffnw(
